pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
]
